[wip] splitting the term-manager
