+++ src/asgard_term_manager.hpp
@@ -1,68 +1,9 @@
 #pragma once
 
-#include "asgard_term_sources.hpp"
+#include "asgard_term_apply.hpp"
 
 namespace asgard
 {
-
-//! \brief Combines a term with data used for linear operations
-template<typename P>
-struct term_entry {
-  //! make default entry, needs to be re-initialized
-  term_entry() = default;
-  //! initialize the entry with the given term
-  term_entry(term_md<P> tin);
-  //! resource (mpi-rank/gpu) that will own this term
-  resource rec;
-  //! the term, moved from the pde definition
-  term_md<P> tmd;
-  //! coefficient matrices for the term
-  std::array<block_sparse_matrix<P>, max_num_dimensions> coeffs;
-  #ifdef ASGARD_USE_GPU
-  //! gpu coefficient matrices for different levels
-  std::array<std::vector<gpu::vector<P>>, max_num_dimensions> gpu_lcoeffs;
-  //! pointers to gpu matrices
-  std::array<gpu::vector<P*>, max_num_dimensions> gpu_coeffs;
-  #endif
-  //! ADI pseudoinverses of the coefficients
-  std::array<block_sparse_matrix<P>, max_num_dimensions> adi;
-  //! if the term has additional mass terms, term 0 will contain the mass-up-to current level
-  std::array<block_diag_matrix<P>, max_num_dimensions> mass;
-  //! kronmult operation permutations
-  kronmult::permutes perm;
-  //! dependencies on the Poisson solver
-  bool has_poisson = false;
-  //! indicates if this a single term or a chain, negative means member of a chain
-  int num_chain = 1;
-  //! left/right boundary conditions source index, if positive
-  int bc_source_id = -1;
-
-  //! check if the 1d term needs a Poisson solver
-  static bool has_needs_poisson(term_1d<P> const &t1d);
-
-  //! boundary conditions, start and end
-  indexrange<int> bc;
-  //! dimension holding a flux, -1 if no flux
-  int flux_dim = -1;
-
-  //! returns true if this is the beginning of a chain with at least one more term
-  bool is_chain_start() const { return (num_chain > 1); }
-  //! returns true if this is a link in a chain, false if stand-alone or first link
-  bool is_chain_link() const { return (num_chain < 0); }
-  //! mark the entry as being part of a chain
-  void mark_as_chain_link() { num_chain = -1; }
-  //! retrun true if the term is separable
-  bool is_separable() const { return (not is_interpolatory); }
-
-  //! indicates whether the term is interpolatory
-  bool is_interpolatory = false;
-  //! interpolation always uses ifield, e.g., first in the chain
-  bool interp_uses_ifield = false;
-  //! interpolation uses the moments or just the field
-  bool interp_uses_moments = false;
-  //! interpolation goes to hierarchical basis only or goes all the way to wavelets
-  bool interp_stop_at_hierarchy = false;
-};
 
 /*!
  * \brief Manages the terms and matrices, also holds the mass-matrices and kronmult-workspace
